# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Address Microservice
# local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /addresses:
    x-swagger-router-controller: address_controller
    get:
      description: Get all addresses in database
      operationId: searchAddresses
      responses:
        "200":
          description: Success
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ShowAddressResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create address in database
      operationId: newAddress
      parameters:
        - name: user
          in: body
          description: The address
          required: true
          schema:
            $ref: "#/definitions/CreateAddressRequest"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/CreateAddressResponse"
        # responses may fall through to errors
        '400':
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /address/{Address_id}:
    x-swagger-router-controller: address_controller
    get:
      description: Read an Address in database
      operationId: getAddress
      parameters:
        - name: Address_id
          in: path
          description: get by address id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateAddressResponse"
        '400':
          description: Invalid ID supplied
        '404':
          description: Person not found
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update an Address in database
      operationId: updateAddress
      parameters:
        - name: Address_id
          in: path
          description: update by address id
          required: true
          type: string
        - name: address
          in: body
          description: The address object used to update db data
          required: true
          schema:
            $ref: "#/definitions/UpdateAddressRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ShowAddressResponse"
        '400':
          description: invalid input, object invalid
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete an Address in database
      operationId: deleteAddress
      parameters:
        - name: Address_id
          in: path
          description: The address id
          required: true
          type: string
      responses:
        "200":
          description: Success
        '400':
          description: invalid input, object invalid
        '409':
          description: an existing item already exists  
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    
# complex objects have schema definitions
definitions:
  ShowAddressResponse:
    required:
      - City
      - Country
      - Persons
      - Address_id
      - State
    properties:
      City:
        type: string
      Country:
        type: string
      Persons:
        type: object
      Address_id:
        type: string
      State:
        type: string
  UpdateAddressRequest:
    properties:
      Persons:
        type: object
      Address_id:
        type: string
      Country:
        type: string
  CreateAddressRequest:
    required:
      - City
      - Country
      - Persons
      - Address_id
      - State
    properties:
      City:
        type: string
      Country:
        type: string
      Persons:
        type: object
      Address_id:
        type: string
      State:
        type: string
  CreateAddressResponse:
    required:
      - Address_id
    properties:
      Address_id:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string